Oisin branch

+++ index.js
@@ -8,14 +8,15 @@ const schedule = require('node-schedule');
 const help = require('./src/commands/help')
 const poll = require('./src/commands/poll')
 const notify = require('./src/commands/notify')
-const accessSpreadsheet = require('./src/commands/accessSpreadsheet') 
+const accessSpreadsheet = require('./src/commands/accessSpreadsheet')
 const addUser = require('./src/commands/addUser')
 const updateUsers = require('./src/commands/updateUsers')
 const userMeritCount = require('./src/commands/userMeritCount')
 const addNewEvent = require('./src/commands/newEvent')
 const attendance = require('./src/commands/attendance')
+const weeklySchedule = require('./src/commands/weeklySchedule')
 
-
+const devChannelID = "693888511010406421";
 
 mongoose.connect(process.env.DATABASE_URL, { dbName: 'test', useNewUrlParser: true })
 
@@ -42,19 +43,26 @@ bot.on('message', msg => {
     censor = censor.toLowerCase();
 
     if (censor.includes("nigger") || censor.includes("nigga")) {
+        if(!msg.author.bot){
+            let user = msg.member;
+            const channelId = devChannelID;
+            const myChan = bot.channels.find(chan => chan.id === channelId);
 
-        let user = msg.member;
-        user.kick("You've been warned");
-        msg.delete(); 
+            if(user.kickable){
+                user.kick("You've been warned")
+                msg.delete();
 
-        const channelId = "501377189703450624";
+                //posts in dev channel with who was banned and the offending message
+                myChan.send("User kicked for use of proscribed word: " + user + "\nMessage was: " + msg.content);
 
-        const myChan = bot.channels.find(chan => chan.id === channelId);
+                console.log("Test word kicked: " + msg.content);
+            }
 
-        myChan.send("User kicked for n word ussage: " + user + "\nMessage was: " + msg.content);
-
-        console.log("Test word kicked: " + msg.content);
-
+            else{
+                myChan.send("User not kicked due to perms/error " + user + "\nMessage was: " + msg.content);
+                console.log("Test word kicked: " + msg.content);
+            }
+        }
     }
 
 
@@ -79,7 +87,7 @@ bot.on('message', msg => {
         msg.channel.send(`Version: ${package.version}`)
     }
 
-    
+
     else if (command === 'help' || command === 'h') {
         help(msg)
     }
@@ -95,7 +103,7 @@ bot.on('message', msg => {
     else if (command === 'update' && true /* admin */) {
         console.log("Command update read in. List update");
         updateUsers(bot, msg, args)
-       
+
     }
 
     else if (command === 'add' && true /* admin */) {
@@ -107,7 +115,7 @@ bot.on('message', msg => {
     else if (command === 'new' && true /* admin */) {
         addUser(bot, msg, args);
         console.log("Command newUser read in.");
-        
+
     }
 
     else if (command === 'merits' ) {
@@ -127,7 +135,14 @@ bot.on('message', msg => {
         console.log("Command event (attendance) read in.");
 
     }
-  
+
+    else if (command === 'schedule') {
+        weeklySchedule(msg);
+        console.log("Command event (schedule) read in.");
+
+
+    }
+
 })
 
 bot.on('error', err => {
@@ -137,4 +152,3 @@ bot.on('error', err => {
 // Invite link
 // https://discordapp.com/oauth2/authorize?client_id=501866111017680911&scope=bot
 bot.login(process.env.DISCORD_ACCESS_TOKEN);
-
